Various Python 3 support updates Add Dockerfile Update README Begin splitting main file out to more manageable chunks Add a couple of new tests
