+++ .gitignore
@@ -1,0 +1,3 @@
+.venv
+__db.*
+__pycache__
